Add createUser and checkIfUserExists functions to db-service
The `db-service` module now includes two new functions: `createUser` and `checkIfUserExists`. These functions allow for creating a new user in the database and checking if a user already exists based on their username and email. This enhances the functionality of the application by providing database operations related to user management.